fix: change finalizeAmunt to [withdrawStreamReturnValue, cancelStreamReturnValues] functions for PR #41
