+++ src/utils/soroban/withdrawStreamReturnValue.ts
@@ -1,0 +1,13 @@
+import { scValToNative, xdr } from 'stellar-sdk';
+import { Api } from 'stellar-sdk/lib/soroban';
+
+const { scvI128 } = xdr.ScVal;
+
+const withdrawStreamReturnValue = (finalize: Api.GetSuccessfulTransactionResponse) => {
+  const withdrawXdr = Object(finalize?.returnValue)._value;
+  const withdrawScVal = scvI128(withdrawXdr);
+
+  return scValToNative(withdrawScVal);
+};
+
+export default withdrawStreamReturnValue;
